@@ -59,8 +59,8 @@
 	"dependencies": {
 		"prop-types": "^15.7.2",
 		"query-string": "^7.0.1",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1",
 		"react-polyfills": "0.0.1",
 		"react-router": "^6.0.2"
 	}
