@@ -1,12 +1,11 @@
-//import react into the bundle
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
+import Layout from "./component/layout/layout";
 
-// include your styles into the webpack bundle
-import "../styles/index.css";
+import "../styles/index.css"
 
-//import your own components
-import Home from "./component/home.jsx";
 
-//render your react application
-ReactDOM.render(<Home />, document.querySelector("#app"));
+
+
+
+createRoot(document.querySelector("#app")).render(<Layout/>)
